+++ benchmark/test.ts
@@ -1,19 +1,43 @@
 import { Batch, Concurrency } from '@lzptec/concurrency';
+import { inspect } from 'util';
 
 const data = Array.from({ length: 15 }, (_, idx) => idx);
 Batch;
 Concurrency;
 
+const globalMaxConcurrency = 4;
+const instanceMaxConcurrency = 4;
 
+const globalBatchSize = 4;
+const instanceBatchSize = 4;
 
 const run = async () => {
     let idx = 0;
     const batchInstance = new Batch({
-        batchSize: 4
+        batchSize: instanceBatchSize
     });
     const concurrencyInstance = new Concurrency({
-        maxConcurrency: 4
+        maxConcurrency: instanceMaxConcurrency
     });
+
+    console.info(`Concurrency Settings: `);
+    console.info(inspect({
+        'Global Max Concurrency': globalMaxConcurrency,
+        'Instance Max Concurrency': instanceMaxConcurrency
+    }, {
+        compact: false,
+        colors: true
+    }));
+
+
+    console.info(`Batch Settings: `);
+    console.info(inspect({
+        'Global Batch Size': globalBatchSize,
+        'Instance Batch Size': instanceBatchSize
+    }, {
+        compact: false,
+        colors: true
+    }));
 
     // const mapSharedBatch = {
     //     'BatchInstance.map 4': await Promise.all([
@@ -42,9 +66,11 @@ const run = async () => {
     // }
     // mapSharedBatch;
 
+    console.log(`-- Map --`);
     const map = {
-        'Batch.map 4': await Batch.map(data, {
-            batchSize: 4,
+        'Batch.map': await Batch.map({
+            input: data,
+            batchSize: globalBatchSize,
             task: async (item) => new Promise((resolve) => {
                 console.log(`Batch.map ${idx++} - Item ${item}`);
                 setTimeout(() => resolve(item * item + 1), 400 + (idx * 50));
@@ -52,14 +78,15 @@ const run = async () => {
         }),
         _: (() => { idx = 0; globalThis.gc?.(); })(),
 
-        'BatchInstance.map 4': await batchInstance.map(data, async (item) => new Promise((resolve) => {
+        'BatchInstance.map': await batchInstance.map(data, async (item) => new Promise((resolve) => {
             console.log(`BatchInstance.map ${idx++} - Item ${item}`)
             setTimeout(() => resolve(item * item + 1), 400 + (idx * 50));
         })),
         _0: (() => { idx = 0; globalThis.gc?.(); })(),
 
-        'Concurrency.map 4': await Concurrency.map(data, {
-            maxConcurrency: 4,
+        'Concurrency.map': await Concurrency.map({
+            input: data,
+            maxConcurrency: globalMaxConcurrency,
             task: async (item) => new Promise((resolve) => {
                 console.log(`Concurrency.map ${idx++}`)
                 setTimeout(() => resolve(item * item + 1), 400 + (idx * 50));
@@ -67,7 +94,7 @@ const run = async () => {
         }),
         _1: (() => { idx = 0; globalThis.gc?.(); })(),
 
-        'ConcurrencyInstance.map 4': await concurrencyInstance.map(data, async (item) => new Promise((resolve) => {
+        'ConcurrencyInstance.map': await concurrencyInstance.map(data, async (item) => new Promise((resolve) => {
             console.log(`ConcurrencyInstance.map ${idx++}`)
             setTimeout(() => resolve(item * item + 1), 400 + (idx * 50));
         })),
@@ -81,11 +108,50 @@ const run = async () => {
     map;
     console.log(JSON.stringify(map));
 
+    console.log(`-- Group --`);
+    const group = {
+        'Batch.group': await Batch.group({
+            input: data,
+            batchSize: globalBatchSize,
+            task: async (item) => new Promise((resolve) => {
+                console.log(`Batch.group ${idx++} - Item ${item}`);
+                setTimeout(() => resolve(item.toString()), 400 + (idx * 50));
+            })
+        }),
+        _: (() => { idx = 0; globalThis.gc?.(); })(),
+
+        'BatchInstance.group': await batchInstance.group(data, async (item) => new Promise((resolve) => {
+            console.log(`BatchInstance.group ${idx++} - Item ${item}`)
+            setTimeout(() => resolve(item.toString()), 400 + (idx * 50));
+        })),
+        _0: (() => { idx = 0; globalThis.gc?.(); })(),
+
+        'Concurrency.group': await Concurrency.group({
+            input: data,
+            maxConcurrency: globalMaxConcurrency,
+            task: async (item) => new Promise((resolve) => {
+                console.log(`Concurrency.group ${idx++}`)
+                setTimeout(() => resolve(item.toString()), 400 + (idx * 50));
+            })
+        }),
+        _1: (() => { idx = 0; globalThis.gc?.(); })(),
+
+        'ConcurrencyInstance.group': await concurrencyInstance.group(data, async (item) => new Promise((resolve) => {
+            console.log(`ConcurrencyInstance.group ${idx++}`)
+            setTimeout(() => resolve(item.toString()), 400 + (idx * 50));
+        })),
+        _2: (() => { idx = 0; globalThis.gc?.(); })(),
+    };
+    group;
+    console.log(JSON.stringify(group));
+
     const filterSymbol = Symbol();
 
+    console.log(`-- Filter --`);
     const filter = {
-        'Batch.filter 4': await Batch.filter(data, {
-            batchSize: 4,
+        'Batch.filter': await Batch.filter({
+            input: data,
+            batchSize: globalBatchSize,
             predicate: async (item) => new Promise((resolve) => {
                 console.log(`Batch.filter ${idx++} - Item ${item}`)
                 setTimeout(() => resolve(item % 2 === 0), 400 + (idx * 50));
@@ -93,14 +159,15 @@ const run = async () => {
         }),
         _: (() => { idx = 0; globalThis.gc?.(); })(),
 
-        'BatchInstance.filter 4': await batchInstance.filter(data, async (item) => new Promise((resolve) => {
+        'BatchInstance.filter': await batchInstance.filter(data, async (item) => new Promise((resolve) => {
             console.log(`BatchInstance.filter ${idx++} - Item ${item}`)
             setTimeout(() => resolve(item % 2 === 0), 400 + (idx * 50));
         })),
         _0: (() => { idx = 0; globalThis.gc?.(); })(),
 
-        'Concurrency.filter 4': await Concurrency.filter(data, {
-            maxConcurrency: 4,
+        'Concurrency.filter': await Concurrency.filter({
+            input: data,
+            maxConcurrency: globalMaxConcurrency,
             predicate: async (item) => new Promise((resolve) => {
                 console.log(`Concurrency.filter ${idx++}`)
                 setTimeout(() => resolve(item % 2 === 0), 400 + (idx * 50));
@@ -108,7 +175,7 @@ const run = async () => {
         }),
         _1: (() => { idx = 0; globalThis.gc?.(); })(),
 
-        'ConcurrencyInstance.filter 4': await concurrencyInstance.filter(data, async (item) => new Promise((resolve) => {
+        'ConcurrencyInstance.filter': await concurrencyInstance.filter(data, async (item) => new Promise((resolve) => {
             console.log(`ConcurrencyInstance.filter ${idx++}`)
             setTimeout(() => resolve(item % 2 === 0), 400 + (idx * 50));
         })),
@@ -122,9 +189,11 @@ const run = async () => {
     filter;
     console.log(JSON.stringify(filter));
 
+    console.log(`-- MapSettled --`);
     const mapSettled = {
-        'Batch.mapSettled 4': await Batch.mapSettled(data, {
-            batchSize: 4,
+        'Batch.mapSettled': await Batch.mapSettled({
+            input: data,
+            batchSize: globalBatchSize,
             task: async (item) => new Promise((resolve) => {
                 console.log(`Batch.mapSettled ${idx++}`)
                 setTimeout(() => resolve(item ** item + 1), 400 + (idx * 50));
@@ -132,8 +201,9 @@ const run = async () => {
         }),
         _: (() => { idx = 0; globalThis.gc?.(); })(),
 
-        'Concurrency.mapSettled 4': await Concurrency.mapSettled(data, {
-            maxConcurrency: 4,
+        'Concurrency.mapSettled': await Concurrency.mapSettled({
+            input: data,
+            maxConcurrency: globalMaxConcurrency,
             task: async (item) => new Promise((resolve) => {
                 console.log(`Concurrency.mapSettled ${idx++}`)
                 setTimeout(() => resolve(item ** item + 1), 400 + (idx * 50));
@@ -151,9 +221,11 @@ const run = async () => {
     mapSettled;
     console.log(JSON.stringify(mapSettled));
 
+    console.log(`-- ForEach --`);
     const forEach = {
-        'Batch.forEach 4': await Batch.forEach(data, {
-            batchSize: 4,
+        'Batch.forEach': await Batch.forEach({
+            input: data,
+            batchSize: globalBatchSize,
             task: async (item) => new Promise<void>((resolve) => {
                 console.log(`Batch.forEach ${idx++}`)
                 setTimeout(() => { item ** item + 1; resolve(); }, 400 + (idx * 50));
@@ -161,8 +233,9 @@ const run = async () => {
         }),
         _: (() => { idx = 0; globalThis.gc?.(); })(),
 
-        'Concurrency.forEach 4': await Concurrency.forEach(data, {
-            maxConcurrency: 4,
+        'Concurrency.forEach': await Concurrency.forEach({
+            input: data,
+            maxConcurrency: globalMaxConcurrency,
             task: async (item) => new Promise<void>((resolve) => {
                 console.log(`Concurrency.forEach ${idx++}`)
                 setTimeout(() => { item ** item + 1; resolve(); }, 400 + (idx * 50));